requirement.txt
acme==2.9.0
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.1.4
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
certbot==2.9.0
certbot-nginx==2.9.0
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
click==8.1.6
cloud-init==24.4
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
ConfigArgParse==1.7
configobj==5.0.8
cryptography==44.0.1
cupshelpers==1.0
dbus-python==1.3.2
debugpy==1.8.6
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
distro-info==1.7+build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==2.1.4
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.8
fasteners==0.18
filelock==3.13.1
greenlet==3.1.1
gyp==0.1
h11==0.14.0
html5lib==1.1
httplib2==0.20.4
idna==3.6
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.2
josepy==1.14.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
jupyter_client==8.6.3
jupyter_core==5.7.2
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mechanize==0.4.9
monotonic==1.6
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
packaging==24.0
paramiko==2.12.0
parsedatetime==2.6
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
platformdirs==4.2.0
ply==3.11
prompt_toolkit==3.0.48
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
PyGObject==3.48.2
PyICU==2.12
PyJWT==2.7.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu3
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==26.2.0
redis==4.3.4
requests==2.31.0
rich==13.7.1
rq==1.16.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.45.3
systemd-python==235
texttable==1.6.7
tomli==2.2.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
userpath==1.9.1
uvicorn==0.34.0
virtualenv==20.25.0+ds
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.7.0
wrapt==1.15.0
xdg==5
xkit==0.0.0
